+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased] - 2017-05-04
 ### Added
 - Added support to get service return value in response middlewares.
+- Added getter for origin duration to transport
 
 ## [1.0.1] - 2017-04-28
 ### Added