+++ CHANGELOG.md
@@ -4,11 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0] - Unreleased
+## [1.1.0] - 2017-06-01
 ### Added
 - Added support to get service return value in response middlewares.
 - Added getter for origin duration to transport.
 - Added "binary" type support for parameters and return value.
+
+### Changed
+- Updated CONTRIBUTING.md and README.md
 
 ## [1.0.1] - 2017-04-28
 ### Added